refactor: 라운드 오디오 오케스트레이션을 훅으로 분리하고 플레이어 타입을 좁힘
## 오케스트레이션 훅 분리

`RoundAudioPlayer`에 6개 `useEffect` + 7개 `useRef` + `loadCurrentTrack`/
`startActivePlayback`/`disableCaptions`가 그대로 남아 있어 `front/CLAUDE.md`의
"비즈니스 로직은 커스텀 훅으로 분리 / 단일 컴포넌트 200줄 초과 시 분리 검토" 대상이었다.

`useRoundAudioOrchestrator`로 명령형 제어를 전부 옮기고
`{ activeIndex, status, playerHandlers(index) }`만 노출한다.
컴포넌트는 308줄 → 49줄이 되어 `ClipPlayer` 두 개를 그리는 일만 한다.

## 플레이어 타입 좁히기

`react-youtube`가 export하는 `YouTubePlayer`는 이 프로젝트에서 쓸 수 없다 —
`youtube-player/dist/types`에서 재export되는데 그 패키지는 선언 파일을 배포하지 않아
(`.js`/`.js.flow`만 존재) tsconfig의 `skipLibCheck: true`가 해석 실패를 삼키고
타입이 조용히 `any`로 떨어진다. 붙여도 오타를 잡지 못한다.

    node_modules/react-youtube/dist/YouTube.d.ts(3,40): error TS7016:
      Could not find a declaration file for module 'youtube-player/dist/types'.

대신 실제 호출하는 표면만 추린 `YouTubePlayerHandle`을 직접 선언했다. 래퍼가 프록시하지
않는 `unloadModule`/`addEventListener`는 옵셔널로 둔다. `ClipPlayer`·`useClipPlayback`·
오케스트레이터의 `any`를 이 타입으로 교체하고 `event.target` 경계 한 곳에서만 단언한다.
메서드명 오타가 컴파일 타임에 잡히는 것을 확인했다:

    error TS2551: Property 'loadVideoByIdd' does not exist on type 'YouTubePlayerHandle'.
    error TS2551: Property 'seekToo' does not exist on type 'YouTubePlayerHandle'.

동작 변경 없음. `npm run typecheck` / `npm run build` 통과.
